# web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Tropicast.Dashboard.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stations/{id}/broadcast-target": {
      "post": {
        "tags": [
          "Desktop"
        ],
        "summary": "Ingest URLs, username and a new password for this desktop device; replaces the device's previous one.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastTargetResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stations/{id}": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "One station; the ETag header is needed to change it.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Version of the resource; send it back as If-Match to change it.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Stations"
        ],
        "summary": "Changes station settings. Needs If-Match.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "The ETag from your last read. Missing: 428; stale: 412.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStationCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Version of the resource; send it back as If-Match to change it.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stations"
        ],
        "summary": "Deletes the station; its history is kept and its public ID is never reused.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "The ETag from your last read. Missing: 428; stale: 412.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioFormat": {
        "enum": [
          "Mp3",
          "Opus"
        ],
        "description": "A stream format, as in the mount `/stations/{id}/live.{mp3,opus}`."
      },
      "BroadcastOutput": {
        "required": [
          "format",
          "contentType",
          "ingestUrl",
          "listenerUrl"
        ],
        "type": "object",
        "properties": {
          "format": {
            "description": "Stream format.",
            "$ref": "#/components/schemas/AudioFormat"
          },
          "contentType": {
            "type": "string",
            "description": "Send it as the source's `Content-Type`."
          },
          "ingestUrl": {
            "type": "string",
            "description": "Where to publish (HTTP PUT, Icecast source protocol, TLS).",
            "format": "uri"
          },
          "listenerUrl": {
            "type": "string",
            "description": "Where listeners play it.",
            "format": "uri"
          }
        },
        "description": "One stream to publish."
      },
      "BroadcastTargetResponse": {
        "required": [
          "stationId",
          "publicId",
          "stationName",
          "credentialId",
          "username",
          "password",
          "maxBitrateKbps",
          "outputs"
        ],
        "type": "object",
        "properties": {
          "stationId": {
            "type": "string",
            "description": "Station ID in the API.",
            "format": "uuid"
          },
          "publicId": {
            "type": "string",
            "description": "Station ID in mounts and listener URLs."
          },
          "stationName": {
            "type": "string",
            "description": "Station name, e.g. for the `Ice-Name` header."
          },
          "credentialId": {
            "type": "string",
            "description": "The device's broadcast credential, listed under the station's credentials.",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "description": "Source username: the station's public ID."
          },
          "password": {
            "type": "string",
            "description": "Source password of this device. Shown only here: store it now. Asking again replaces it."
          },
          "maxBitrateKbps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Highest bitrate the plan accepts, per stream.",
            "format": "int32"
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BroadcastOutput"
            },
            "description": "Streams the plan allows, MP3 first."
          }
        },
        "description": "Everything the desktop app needs to go live on a station. Contract version 1 (docs/desktop-api.md); it never\ncontains node, admin or relay credentials."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ListenerUrls": {
        "required": [
          "mp3",
          "opus"
        ],
        "type": "object",
        "properties": {
          "mp3": {
            "type": "string",
            "description": "MP3 stream, playable everywhere.",
            "format": "uri"
          },
          "opus": {
            "type": [
              "null",
              "string"
            ],
            "description": "Ogg Opus stream: about half the mobile data of MP3.",
            "format": "uri"
          }
        },
        "description": "Listener URLs of a station. Uri? ListenerUrls.Opus is null when the plan has no Opus."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StationResponse": {
        "examples": [
          {
            "id": "01927f5e-6c1a-7b3e-9a52-3f1d2c4b5a69",
            "publicId": "k3m9x2p7qa",
            "name": "Radio Mada",
            "slug": "radio-mada",
            "description": "News and salegy from Antananarivo",
            "genre": "Talk",
            "country": "MG",
            "language": "mg",
            "logoUrl": null,
            "website": "https://radiomada.example",
            "listInDirectory": true,
            "createdAt": "2026-10-09T12:00:00+00:00",
            "listenerUrls": {
              "mp3": "https://listen.tropicastradio.com/stations/k3m9x2p7qa/live.mp3",
              "opus": "https://listen.tropicastradio.com/stations/k3m9x2p7qa/live.opus"
            }
          }
        ],
        "required": [
          "id",
          "publicId",
          "name",
          "slug",
          "description",
          "genre",
          "country",
          "language",
          "logoUrl",
          "website",
          "listInDirectory",
          "createdAt",
          "listenerUrls"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID in the API.",
            "format": "uuid"
          },
          "publicId": {
            "type": "string",
            "description": "ID in listener URLs and Icecast mounts; never changes."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "slug": {
            "type": "string",
            "description": "URL name, unique in the tenant."
          },
          "description": {
            "type": "string",
            "description": "Shown on the public page and in directories."
          },
          "genre": {
            "type": "string",
            "description": "Music or programme genre."
          },
          "country": {
            "type": [
              "null",
              "string"
            ],
            "description": "ISO 3166-1 alpha-2 code."
          },
          "language": {
            "type": [
              "null",
              "string"
            ],
            "description": "BCP 47 language tag."
          },
          "logoUrl": {
            "type": [
              "null",
              "string"
            ],
            "description": "Logo address.",
            "format": "uri"
          },
          "website": {
            "type": [
              "null",
              "string"
            ],
            "description": "The station's website.",
            "format": "uri"
          },
          "listInDirectory": {
            "type": "boolean",
            "description": "Listed on RadioBrowser (only when the plan includes it)."
          },
          "createdAt": {
            "type": "string",
            "description": "When the station was created.",
            "format": "date-time"
          },
          "listenerUrls": {
            "description": "Where listeners play the station.",
            "$ref": "#/components/schemas/ListenerUrls"
          }
        },
        "description": "A station."
      },
...
      "UpdateStationCommand": {
        "examples": [
          {
            "genre": "Salegy",
            "description": "Music all night"
          }
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "slug": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "genre": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "type": [
              "null",
              "string"
            ]
          },
          "language": {
            "type": [
              "null",
              "string"
            ]
          },
          "logoUrl": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "website": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "listInDirectory": {
            "type": [
              "null",
              "boolean"
            ]
          }
        },
        "description": "Changes station settings; omitted (null) fields keep their value."
      },
...
    }
  },
...
}